requirements.txt
pytest~=6.2.4
selenium~=3.141.0
allure-pytest~=2.9.43
requests
cerberus
deepdiff
jsonschema
faker
